Specify --target-port on docs and match port on quickstart guide
AFAIK we have 2 ports: the proxy port in which ToolHive runs and
a hardcoded port that may be used by an MCP server. For the second
one, the current flag should be `--target-port` instead of `--port`.
Changed the only reference I could find in docs.

Also, there was a small typo in the quickstart guide in the ports.
The text was talking about port `49226` but the port shown in the
example was port `15266`.